nop bai tap ngay

+++ module4/ss9_AOP/bai_tap/src/main/java/com/codegym/muon_sach/service/IDetailBookService.java
@@ -14,5 +14,5 @@ public interface IDetailBookService {
 
     void deleteByCode(Long code);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 }
